image_service/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
psycopg2-binary==2.9.7
sqlalchemy==2.0.21
google-generativeai==0.3.1
python-dotenv==1.0.0
python-multipart==0.0.6
pillow==10.0.1
redis==5.0.0
pydantic==2.4.2
httpx==0.25.0 